Add financial institution name to balance history csv files
The account data from Mint includes a "fiName" that contains the financial institution's name. This change maps that field and includes it at the end of the resulting file names. This can be incredibly helpful while importing the blance histories into their respective accounts when account names are very similar, or even identical.